rearrange file names
