+++ src/shared/shared_ck/particle_dynamics/fluid_dynamics/boundary_condition/all_fluid_boundary_condition_ck.h
@@ -21,22 +21,18 @@
  *                                                                           *
  * ------------------------------------------------------------------------- */
 /**
- * @file    all_shared_fluid_dynamics_ck.h
+ * @file    all_fluid_boundary_condition_ck.h
  * @brief   Head file for all shared physics dynamics for both 2- and 3D build.
  *          This is the header file that user code should include to pick up all
             particle dynamics capabilities.
  * @author	Chi Zhang and Xiangyu Hu
  */
 
-#ifndef ALL_SHARED_FLUID_DYNAMICS_CK_H
-#define ALL_SHARED_FLUID_DYNAMICS_CK_H
+#ifndef ALL_FLUID_BOUNDARY_CONDITION_CK_H
+#define ALL_FLUID_BOUNDARY_CONDITION_CK_H
 
-#include "acoustic_step_1st_half.hpp"
-#include "acoustic_step_2nd_half.hpp"
-#include "density_regularization.hpp"
-#include "all_fluid_boundary_condition_ck.h"
-#include "fluid_time_step_ck.hpp"
-#include "transport_velocity_correction_ck.hpp"
-#include "viscous_force.hpp"
+#include "fluid_boundary_ck.hpp"
+#include "pressure_boundary_ck.hpp"
+#include "velocity_boundary_ck.hpp"
 
-#endif // ALL_SHARED_FLUID_DYNAMICS_CK_H
+#endif // ALL_FLUID_BOUNDARY_CONDITION_CK_H
